Refactored DAOs (New Method to create dynamic Select statements) Some classes are marked with @NotNull Annotations, these do not check the parameter if they are null Added some Classes for the Statistical stuff Refactored TableCells -> new parent class and comments Added some new TableCell classes to the components package Extracted BoxUIWrapper from Controller Some controllers need more logging -> started with added logger to controller Refactored some Box Tests Added @NotNull Annotations to Service layer

+++ src/sepm/ss17/e1526280/gui/dialogs/DialogUtil.java
@@ -7,6 +7,8 @@ import org.slf4j.LoggerFactory;
 import sepm.ss17.e1526280.util.DatabaseService;
 import sepm.ss17.e1526280.util.GlobalSettings;
 
+import java.util.concurrent.atomic.AtomicBoolean;
+
 /**
  * A simple Utility class which can spawn different dialogs with pre-defined Text messages
  *
@@ -17,6 +19,7 @@ public class DialogUtil {
 
     /** Logger for logging ... duh **/
     private static final Logger LOG = LoggerFactory.getLogger(DialogUtil.class);
+    private static final AtomicBoolean lock = new AtomicBoolean(true);
 
     /**
      * Displays an ExceptionAlert Dialog
@@ -40,24 +43,27 @@ public class DialogUtil {
 
     /**
      * Displays an Exception Dialog and waits for the Dialog to close, after
-     * that the Application shutdown is encored
+     * that the Application shutdown is forced, only one dialog may be created
      * @param err the Exception which should be displayed
      * @return null
      */
     public static Void onFatal(Throwable err) {
         LOG.error("There was a Fatal Error: " + err.getMessage());
 
-        Platform.runLater(() -> {
-            Alert alert = new ExceptionAlert(err);
+        if( lock.getAndSet(false) )
+            Platform.runLater(() -> {
+                Alert alert = new ExceptionAlert(err);
 
-            alert.setTitle("Fatal Error: " + GlobalSettings.APP_TITLE);
-            alert.setHeaderText("Es ist ein Fehler aufgetreten:");
-            alert.setContentText("Das Program muss beendet werden");
-            alert.showAndWait();
+                alert.setTitle("Fatal Error: " + GlobalSettings.APP_TITLE);
+                alert.setHeaderText("Es ist ein Fehler aufgetreten:");
+                alert.setContentText("Das Program muss beendet werden");
+                alert.showAndWait();
 
-            DatabaseService.destroyService();
-            Platform.exit();
-        });
+                DatabaseService.destroyService();
+                Platform.exit();
+            });
+        else
+            LOG.warn("Tried to spawn multiple fatal dialogs!");
 
         return null;
     }
